requirements.txt
tensorflow
scikit-learn
Pillow
pandas
jupyter
nbconvert
nbformat
matplotlib
numpy
joblib
torch
psutil
GPUtil
pydantic
fastapi